feat: Incorporación de funciones de seguimiento financiero con gráficos e historial de transacciones.
- Se añadió la biblioteca de gráficos para la visualización de datos.
- Se implementó GastosGraficoCard para mostrar los gastos por categoría mediante un gráfico circular.
- Se creó UltimasTransaccionesCard para mostrar las últimas transacciones con etiquetas de categoría.
- Se mejoró el componente de la aplicación para obtener y calcular ingresos y gastos por categoría.
- Se actualizaron las BalanceCards para reflejar los ingresos y gastos dinámicos.
- Se mejoró el estilo de la interfaz de usuario para una mejor legibilidad y experiencia de usuario.

+++ server/db/database.js
@@ -17,7 +17,7 @@ db.exec(`
     CREATE TABLE IF NOT EXISTS transacciones (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         tipo_id INTEGER NOT NULL,
-        cantidad REAL NOT NULL,
+        cantidad INTEGER NOT NULL,
         categoria_id INTEGER NOT NULL,
         fecha TEXT NOT NULL,
         descripcion TEXT,
